screen-capture-server/requirements.txt
mss>=9.0.0
websockets>=12.0
Pillow>=10.0.0
